Move normalizeSlashes() to Util\Functions class
Follow up on 5

When trying to run the tests on Windows, I ran into more issues with mixed slashes.

As that means that the `normalizeSlashes()` functionality is needed in more places, I'm proposing to introduce a generic `Util\Functions` class for simple static methods which are needed in multiple places.

Includes introducing dedicated tests for the `normalizeSlashes()` method.